Fix turning the screen on/off on some RPI systems
On some systems, vcgencmd is not on PATH. There doesn't seem to be a
simple way to customize env vars passed to Kodi Python scripts but
adding it to the PATH in screensaver.py works.

This should not break anything on other systems, just fix the problem on
OSMC and other similar distributions.